+++ src/pyappm/pyappm_app_init.py
@@ -41,6 +41,8 @@
 import os
 from pathlib import Path
 from datetime import datetime
+import sys
+import subprocess
 
 from pyappm_configuration import PyAPPMConfiguration  # type: ignore
 
@@ -54,6 +56,29 @@ from dotdict import DotDict  # type: ignore
 from pyapp_toml import CreateAppToml  # type: ignore
 
 from pyappm_tools import run_command  # type: ignore
+
+
+def check_dependencies():
+    try:
+        # Check if pip3 is installed
+        subprocess.run(
+            ["pip3", "--version"],
+            check=True,
+            stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE,
+        )
+        # Check if venv is installed
+        subprocess.run(
+            ["python3", "-m", "venv", "--version"],
+            check=True,
+            stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE,
+        )
+    except subprocess.CalledProcessError:
+        print("Please install pip3 and venv to continue.")
+        print("To install pip3, run: sudo apt install python3-pip")
+        print("To install venv, run: sudo apt install python3-venv")
+        sys.exit(404)
 
 
 def write_pyapp_py(path: Path, app_name: str, config: PyAPPMConfiguration) -> None:
@@ -84,6 +109,7 @@ if __name__ == "__main__":
 
 def init_pyapp(path: str, is_service: bool, config: PyAPPMConfiguration) -> None:
     """Initialize the application in the specified directory."""
+    check_dependencies()
     EnsureVirtualEnvIsNotActive()
     pth: Path
     if path == ".":