+++ CHANGELOG.md
@@ -24,3 +24,4 @@ before:
 - Bugfix, fixed path for repository file to expand user.
 - Bugfix, PyAPPMRepositoryManager, init now correctly loads repositories from file without complaining about repositories already existing.
 - Removed repositories as a parameter in init of PyAPPMRepositoryManager.
+- Added dependency check to app init, it now checks for pip3 and venv, if not it tells the user to install them and exits with error 404.